Limit number of sessions returned by server
This is a temporary (I swear!) workaround for bad performance in the client.

When we have lots of sessions in the client, editing the description or tags of
a session is really slow due to the way that we are storing sessions on the
client. Obviously, we should fix this.

However, this workaround just places a limit on the number of sessions the
client will need to worry about, so this should make the performance "good
enough".

Once we have better performance on the client, we probably also want some kind
of pagnation or infinte scroll.