Add Dockerfile for development image and instructions
This is not really something intended for other people to use, so the
Dockerfile is not very secure – in particular, it fetches a bunch of
stuff from untrusted sources and runs it as root.

When I'm working in different environments I find it useful to have a
standardized Docker container for doing development – this is
especially useful on Windows.