Use Sass to compile SCSS to CSS
This should make it easier to improve the CSS in the future.

This required updating the production Dockerfile to contain an
additional step: We now first compile the SCSS in one stage, then
compile the Elm in another stage, and then finally merge those two
outputs into the third stage, where we run the Elixir server.

It's a little complicated, but I'm not sure if there's a way to make it
simpler, or if this is pretty much what multi-stage builds are designed
for.