Make it possible to log in as any user
We still don't actually manage user accounts, but we do now have basic
(totally insecure) handling of sessions.

I'm not confident that the code I've written is idiomatic Elixir, but it
does seem to work. We have an agent keep track of all known sessions,
and then we talk to this agent to figure out if people have valid
session tokens or not.