Fix bug where added tag would not be selected automatically
We added this functionality in the last commit, 0ff5f02 (Automatically
select content when adding tag), but it looks like using 'setTimeout'
doesn't always work. This seems to happen with more sessions, so I
guess it's related to the time that Elm takes to rerender the DOM. We
now use 'requestAnimationFrame' instead, which seems to work fine. Of
course, this is a hack in the first place.