Redirect on login rather than serving POST request
Previously we would just respond with an HTML page to the POST request after a
successful login. This would create awkward behavior; for example, if the user
refreshes the page, the POST request would be sent (usually after the browser
asks for confirmation). Now we respond with a '303 See Other', which makes more
sense.